@@ -5,33 +5,35 @@
   "version": "0.1.0",
   "author": "Dylan Steck <dylancsteck@gmail.com>",
   "dependencies": {
-    "gatsby": "^2.32.11",
-    "gatsby-cli": "^2.19.3",
-    "gatsby-image": "^2.11.0",
+    "gatsby": "^3.9.1",
+    "gatsby-cli": "^3.9.0",
+    "gatsby-image": "^3.9.0",
     "gatsby-plugin-google-gtag": "^3.6.0",
-    "gatsby-plugin-manifest": "^2.12.1",
-    "gatsby-plugin-offline": "^3.10.2",
-    "gatsby-plugin-react-helmet": "^3.10.0",
+    "gatsby-plugin-manifest": "^3.9.0",
+    "gatsby-plugin-offline": "^4.9.0",
+    "gatsby-plugin-react-helmet": "^4.9.0",
     "gatsby-plugin-react-svg": "^3.0.0",
-    "gatsby-plugin-sharp": "^2.14.3",
+    "gatsby-plugin-sharp": "^3.9.0",
     "gatsby-plugin-smoothscroll": "^1.2.0",
     "gatsby-plugin-svg-sprite": "^2.0.1",
-    "gatsby-source-filesystem": "^2.11.1",
-    "gatsby-source-wordpress": "^4.0.0",
-    "gatsby-transformer-sharp": "^2.12.1",
-    "html-react-parser": "^0.13.0",
+    "gatsby-source-filesystem": "^3.9.0",
+    "gatsby-source-wordpress": "^5.9.2",
+    "gatsby-transformer-sharp": "^3.9.0",
+    "html-react-parser": "^1.2.7",
     "jquery": "^3.6.0",
     "moment": "^2.29.1",
     "npm": "^7.7.4",
+    "postcss": "^8.3.5",
+    "postcss-flexbugs-fixes": "^5.0.2",
+    "postcss-loader": "^6.1.1",
     "prop-types": "^15.7.2",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-helmet": "^6.1.0"
   },
   "devDependencies": {
-    "@prisma/cli": "^2.19.0",
     "fslightbox-react": "^1.6.2",
-    "prettier": "2.0.5"
+    "prettier": "^2.3.2"
   },
   "keywords": [
     "gatsby"
